refactor(stores): co-locate module-local stores with their modules (task 2d)
Move collections, history, variables, and settings stores to their respective modules:
- app/src/stores/collections-store.ts -> app/src/modules/collections/collections-store.ts
- app/src/stores/history-store.ts -> app/src/modules/history/history-store.ts
- app/src/stores/variables-store.ts -> app/src/modules/variables/variables-store.ts
- app/src/stores/settings-store.ts -> app/src/modules/settings/settings-store.ts

Update all import paths across the codebase to reference the new module locations.
Remove the four stores from the app/src/stores/index.ts barrel export.

Cross-module consumers (request-builder, hooks) import directly from the new module paths.
