+++ app/src/modules/collections/CollectionTree.tsx
@@ -7,8 +7,8 @@
 
 import { useState, useRef, useEffect, useCallback, useMemo } from "react";
 import { Folder, Plus, Trash2, Edit2, Copy, FolderPlus, Loader2, Download } from "lucide-react";
-import { useNavigationStore, useCollectionsStore, useSaveStore } from "@/stores";
-import { useImportModalStore } from "@/stores";
+import { useNavigationStore, useSaveStore, useImportModalStore } from "@/stores";
+import { useCollectionsStore } from "@/modules/collections/collections-store";
 import {
 	useCollectionsQuery,
 	useMultipleCollectionRequests,
