+++ app/src/stores/index.ts
@@ -10,12 +10,8 @@ export { useNavigationStore, type NavigationContext } from "./navigation";
 export { useEngineStore } from "./engine-store";
 
 // Domain stores
-export { useCollectionsStore } from "./collections-store";
 export { useDashboardStore } from "./dashboard-store";
-export { useHistoryStore } from "./history-store";
 export { useImportModalStore } from "./import-modal-store";
 export { useSessionStore } from "./session-store";
-export { useVariablesStore, type VariableCategory } from "./variables-store";
 export { useSaveStore, type SaveStatus } from "./save-store";
 export { useResponseStore, type StoredResponse } from "./response-store";
-export { useSettingsStore } from "./settings-store";
